// ui/src/app/models/club.model.ts
export class Club {
    clubId?: number;
    clubName?: string;
    clubLocation?: string;
    clubManger?: string;
    clubOwner?: string;
    noOfMatchesPlayed?: number;
    noOfMatchesWon?: number;
    noOfDraws?: number;
    noOfLosses?: number;
    noOfGoalsScored?: number;
    noOfGoalsAgainst?: number;
    goalDifference?: number;
    noOfPointScored?: number;
}